requirements.txt
selenium==4.23.1
pydantic==2.8.2
python-dotenv==1.0.1
rich==13.7.1